Renaming cudnn_conv_rewriter to gpu_conv_rewriter
